+++ src/FilePack/Controller.php
@@ -10,6 +10,7 @@ namespace aryelgois\BankInterchange\FilePack;
 use aryelgois\Utils\Utils;
 use aryelgois\Medools;
 use aryelgois\BankInterchange;
+use aryelgois\MedoolsRouter\Resource;
 
 /**
  * Wrapper controller that can make one or more files
@@ -42,13 +43,68 @@ abstract class Controller
     protected $views;
 
     /**
-     * Generates the View object from a Model
+     * Creates a new FilePack Controller object
      *
-     * @param Medools\Model $model A MODEL_CLASS instance
-     *
-     * @return ViewInterface
+     * @throws \LogicException If MODEL_CLASS is not subclass of Medools\Model
      */
-    abstract protected function getView(Medools\Model $model);
+    public function __construct()
+    {
+        if (!is_subclass_of(static::MODEL_CLASS, Medools\Model::class)) {
+            $message = 'Invalid MODEL_CLASS in ' . static::class;
+            throw new \LogicException($message);
+        }
+    }
+
+    /**
+     * Returns generated views
+     *
+     * Useful to combine multiple FilePack objects into a single .zip
+     *
+     * @return array[]
+     */
+    public function dump()
+    {
+        return $this->views;
+    }
+
+    /**
+     * Creates a new instance with models in a Resource, then outputs it
+     *
+     * @param Resource $resource Processed MedoolsRouter Resource
+     *
+     * @return boolean For success or failure
+     *
+     * @throws \LogicException If $resource->model_class is invalid
+     */
+    public static function fromResource(Resource $resource)
+    {
+        if (!is_subclass_of($resource->model_class, Medools\Model::class)) {
+            $message = 'Invalid Resource model_class in ' . static::class . ': '
+                . $resource->model_class;
+            throw new \LogicException($message);
+        }
+
+        $controller = new static;
+
+        $list = $resource->getList();
+        if (empty($list)) {
+            return false;
+        }
+
+        foreach ($list as $id) {
+            if (!$controller->generate($id)) {
+                return false;
+            }
+        }
+
+        if ($resource->content_type === 'application/zip') {
+            $controller->zip();
+        } else {
+            $controller->output();
+        }
+
+        return true;
+    }
 
     /**
      * Generates the file View from data in a Medools Model
@@ -57,18 +113,10 @@ abstract class Controller
      * @param string $name  Name for the generated file
      *
      * @return boolean For success or failure
-     *
-     * @throws \LogicException If MODEL_CLASS is not subclass of Medools\Model
      */
     public function generate($where, string $name = null)
     {
-        $model_class = static::MODEL_CLASS;
-        if (!is_subclass_of($model_class, Medools\Model::class)) {
-            $message = 'Invalid MODEL_CLASS in ' . static::class;
-            throw new \LogicException($message);
-        }
-
-        $model = $model_class::getInstance($where);
+        $model = (static::MODEL_CLASS)::getInstance($where);
         if ($model === null) {
             return false;
         }
@@ -85,6 +133,15 @@ abstract class Controller
 
         return true;
     }
+
+    /**
+     * Generates the View object from a Model
+     *
+     * @param Medools\Model $model A MODEL_CLASS instance
+     *
+     * @return ViewInterface
+     */
+    abstract protected function getView(Medools\Model $model);
 
     /**
      * Outputs the View with appropriated headers