+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Setup
 - Dependencies:
   - [aryelgois/databases]
+  - [aryelgois/medools-router]
   - [aryelgois/yasql-php]
   - [symfony/yaml]
   - Platform requirements:
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Config files:
   - Builder config for [aryelgois/yasql-php]
   - Return file parser options for Banco do Nordeste's CNAB400 schema
+  - Router config
 - Database:
   - Populate `wallets` for Banco do Nordeste
   - Column `billet` in `currency_codes`
@@ -43,6 +45,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Column `emission` in `titles`
   - Columns for `protest` in `titles`
   - SQL Programs
+  - Authentication database
 - Namespace `aryelgois\BankInterchange\ReturnFile`
 - Utils:
   - `addExtension()`
@@ -65,6 +68,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ShippingFile `date()`
 - Optional 'R' segment in CNAB240 ShippingFile
 - ShippingFile movement masks
+- `public/`
 
 ### Changed
 - Update dependencies
@@ -176,6 +180,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [aryelgois/databases]: https://github.com/aryelgois/databases
 [aryelgois/Medools]: https://github.com/aryelgois/Medools
+[aryelgois/medools-router]: https://github.com/aryelgois/medools-router
 [aryelgois/utils]: https://github.com/aryelgois/utils
 [aryelgois/yasql]: https://github.com/aryelgois/yasql
 [aryelgois/yasql-php]: https://github.com/aryelgois/yasql-php